Support Pinboard as well as Delicious
The only weirdness here is that tags that have been entered separated by
spaces in Pinboard come out as one tag, e.g. what you get back for
'tags' in a feedparse entry might be:

 [{'label': None, 'scheme': None, 'term': u'pop-culture racism sexism'}]

FIXME another time, since the only effect on presentation is not getting
that tags comma-separated.